requirements.txt
Adafruit-PlatformDetect==3.75.0
adafruit-python-shell==1.8.1
args==0.1.0
cffi==1.17.1
clint==0.5.1
numpy==2.1.3
paho-mqtt==2.1.0
pycparser==2.22
sounddevice==0.5.1
